PLAN-515 PLAN-575 schedule approval and wiring
Now there is a schedule store and a schedule status mixin that controls visibility etc
of everything necessary. Still needs to be wired to the backend.

Schedule approval is less wired up - no store yet because i don't know where that's coming from
or going to. I suspect it will be on the person, not going to bother writing a store for now.