Improve nav layout on small screens
Bootstrap has a breakpoint at 576px. On screens smaller than 576px
columns are switched to be rows, so that they are laid out vertically
rather than horizontally.

Because `#side-navbar .nav` was always calculating it's hight as 100vh -
footer, regardless of the breakpoint, when the bootstrap layout switch
to a columnular view, the nav bar was taking up the whole screen, and
the tab content was pushed below it. That made it look broken, and like
the tabs hadn't loaded at all. This has led to several people thinking
the state of planorama on mobile is even worse than it actually is.

This patch changes to only calculate the height if we are wider than the
576px breakpoint. If we are not, and hence in the columnular view, we
just let the height be calculated from the nav content, meaning we no
longer push the tag content down off screen.

There is still a lot of work to do to make the UI fully functional on
small screens, but this at least makes it so it doesn't appear to be
completely unusable.