Improve header layout on small screens
The header has a fixed height of 100px. This meant that, as the content
wrapped on small screens, the content was bleeding outside of the header
element to under the navigation. Because the navigation is over the top,
you then can't click the logout button.

I'm not sure why the fixed height is needed. But, just in case there is
a good reason for it in the larger screen sizes, I have made a
conservitive revision and only remove the height on smaller screen
sizes.

Technically we could switch at a smaller screen size, but the 576px
width is used by bootstrap as a breakpoint, and is close enough to the
actual value that it seems to make sense to use it here.