Scroll to the top of survey pages
When you click next/previous on a survey page, it doesn't scroll to the
top. This makes it confusing and difficult for people to fill out the
survey, and necessitates a whole paragraph in the introductory text
warning people about the behaviour.

This patch makes it so that clicking next/previous will scroll to the
top of the survey page.

It requires a bit of a hack, because we're not actually using the body
to scroll, instead using a nested scroll container. This means we can't
relay on Vue's scrollBehaviour, and instead have to find and scroll the
container ourselves.

Because I am not too familiar with the codebase, I have limited the
scope of the change to just the surveys, which I know is a big problem
and can easily test. The behaviour can be added to other pages by adding
the `main-content` id.

Long term, the real fix is to stop using a nested scroll container and
instead rely on the body to scroll. This would also improve
accessibility, making it easier for assistive devices to scroll the
page, and allow you to scroll the page with the mousewheel while the
cursor is on the sidebar.

In order to pull that off, we would need to make the header, sidebar and
footer sticky instead. However, because of the complicated DOM and CSS
structure already in place, that will require quite a bit of surgery to
make it work.